requirements.txt
grvt-pysdk>=0.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
requests>=2.31.0
x10-python-trading-starknet>=1.0.0
websockets>=12.0
curl_cffi>=0.15.0
